complete ToString and FromString in Traits
